+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 
 ### Initial Release
 
+## 0.0.10 (2020-09-20)
+* Re-enabled power status polling (was turned off some time back)
+* Adapted current input to keep reverting back to 999 no input, so as to clear the Tile display
+
+
 ## 0.0.9 (2020-09-20)
 * Added support of default wait(). If no wait time is specified, it defaults to 100ms
 * Added support of multi-key without wait(). If no wait is specified, a default wait of 100ms is used